Restructure index tests to avoid running redundant tests
